Fix pkgconfig.configure_extension() breakage
The [pkgconfig.configure_extension()](ihttps://pypi.org/project/pkgconfig/) is quite convenient except when it isn't.
On Deb12.2.0-14 it was adding an empty string to the extra_compile_args:

    extension.extra_compile_args = [ '' ]

Which tricked the system into a command that looked like (irrelevant
info stripped):

    /path/to/C-compiler [options] -o build/.../_ppgplot.o ""

Note the literal'""' at the end of the command; it resulted in the GCC
compiler deciding "Not performing linking because linker command file
cannot be found" (or words to that effect).

Stripping the empty string fixed it