Fixes for local build [TODO XXX after new PyPI!]
In order to have a local recipe build succeed, needed to skip the
LICENSE file b/c the source is pulled from the official PyPI
distribution and that one hasn't got the LICENSE in (yet).

So we add it to MANIFEST.in so by the next PyPI release it will be in
the distro, and then we can modify the recipe back to include the
license file.